+++ x/lockup/keeper/lock.go
@@ -2,7 +2,6 @@ package keeper
 
 import (
 	"fmt"
-	"sort"
 	"time"
 
 	"github.com/cosmos/cosmos-sdk/store/prefix"
@@ -301,9 +300,9 @@ func (k Keeper) addTokensToLock(ctx sdk.Context, lock *types.PeriodLock, coins s
 	}
 
 	// modifications to accumulation store
-	for _, coin := range coins {
-		k.accumulationStore(ctx, coin.Denom).Increase(accumulationKey(lock.Duration), coin.Amount)
-	}
+	// for _, coin := range coins {
+	// 	k.accumulationStore(ctx, coin.Denom).Increase(accumulationKey(lock.Duration), coin.Amount)
+	// }
 
 	return nil
 }
@@ -370,13 +369,14 @@ func (k Keeper) ClearAccumulationStores(ctx sdk.Context) {
 // ResetAllLocks takes a set of locks, and initializes state to be storing
 // them all correctly. This utilizes batch optimizations to improve efficiency,
 // as this becomes a bottleneck at chain initialization & upgrades.
+// UPDATE: accumulation store is disabled in the v4 upgrade.
 func (k Keeper) ResetAllLocks(ctx sdk.Context, locks []types.PeriodLock) error {
 	// index by coin.Denom, them duration -> amt
 	// We accumulate the accumulation store entries separately,
 	// to avoid hitting the myriad of slowdowns in the SDK iterator creation process.
 	// We then save these once to the accumulation store at the end.
-	accumulationStoreEntries := make(map[string]map[time.Duration]sdk.Int)
-	denoms := []string{}
+	// accumulationStoreEntries := make(map[string]map[time.Duration]sdk.Int)
+	// denoms := []string{}
 	for i, lock := range locks {
 		if i%25000 == 0 {
 			msg := fmt.Sprintf("Reset %d lock refs, cur lock ID %d", i, lock.ID)
@@ -388,44 +388,44 @@ func (k Keeper) ResetAllLocks(ctx sdk.Context, locks []types.PeriodLock) error {
 		}
 
 		// Add to the accumlation store cache
-		for _, coin := range lock.Coins {
-			// update or create the new map from duration -> Int for this denom.
-			var curDurationMap map[time.Duration]sdk.Int
-			if durationMap, ok := accumulationStoreEntries[coin.Denom]; ok {
-				curDurationMap = durationMap
-				// update or create new amount in the duration map
-				newAmt := coin.Amount
-				if curAmt, ok := durationMap[lock.Duration]; ok {
-					newAmt = newAmt.Add(curAmt)
-				}
-				curDurationMap[lock.Duration] = newAmt
-			} else {
-				denoms = append(denoms, coin.Denom)
-				curDurationMap = map[time.Duration]sdk.Int{lock.Duration: coin.Amount}
-			}
-			accumulationStoreEntries[coin.Denom] = curDurationMap
-		}
+		// for _, coin := range lock.Coins {
+		// 	// update or create the new map from duration -> Int for this denom.
+		// 	var curDurationMap map[time.Duration]sdk.Int
+		// 	if durationMap, ok := accumulationStoreEntries[coin.Denom]; ok {
+		// 		curDurationMap = durationMap
+		// 		// update or create new amount in the duration map
+		// 		newAmt := coin.Amount
+		// 		if curAmt, ok := durationMap[lock.Duration]; ok {
+		// 			newAmt = newAmt.Add(curAmt)
+		// 		}
+		// 		curDurationMap[lock.Duration] = newAmt
+		// 	} else {
+		// 		denoms = append(denoms, coin.Denom)
+		// 		curDurationMap = map[time.Duration]sdk.Int{lock.Duration: coin.Amount}
+		// 	}
+		// 	accumulationStoreEntries[coin.Denom] = curDurationMap
+		// }
 	}
 
 	// deterministically iterate over durationMap cache.
-	sort.Strings(denoms)
-	for _, denom := range denoms {
-		curDurationMap := accumulationStoreEntries[denom]
-		durations := make([]time.Duration, 0, len(curDurationMap))
-		for duration, _ := range curDurationMap {
-			durations = append(durations, duration)
-		}
-		sort.Slice(durations, func(i, j int) bool { return durations[i] < durations[j] })
-		// now that we have a sorted list of durations for this denom,
-		// add them all to accumulation store
-		msg := fmt.Sprintf("Setting accumulation entries for locks for %s, there are %d distinct durations",
-			denom, len(durations))
-		ctx.Logger().Info(msg)
-		for _, d := range durations {
-			amt := curDurationMap[d]
-			k.accumulationStore(ctx, denom).Increase(accumulationKey(d), amt)
-		}
-	}
+	// sort.Strings(denoms)
+	// for _, denom := range denoms {
+	// 	curDurationMap := accumulationStoreEntries[denom]
+	// 	durations := make([]time.Duration, 0, len(curDurationMap))
+	// 	for duration, _ := range curDurationMap {
+	// 		durations = append(durations, duration)
+	// 	}
+	// 	sort.Slice(durations, func(i, j int) bool { return durations[i] < durations[j] })
+	// 	// now that we have a sorted list of durations for this denom,
+	// 	// add them all to accumulation store
+	// 	msg := fmt.Sprintf("Setting accumulation entries for locks for %s, there are %d distinct durations",
+	// 		denom, len(durations))
+	// 	ctx.Logger().Info(msg)
+	// 	for _, d := range durations {
+	// 		amt := curDurationMap[d]
+	// 		k.accumulationStore(ctx, denom).Increase(accumulationKey(d), amt)
+	// 	}
+	// }
 
 	return nil
 }
@@ -482,9 +482,10 @@ func (k Keeper) Lock(ctx sdk.Context, lock types.PeriodLock) error {
 	}
 
 	// add to accumulation store
-	for _, coin := range lock.Coins {
-		k.accumulationStore(ctx, coin.Denom).Increase(accumulationKey(lock.Duration), coin.Amount)
-	}
+	// disabled in v4
+	// for _, coin := range lock.Coins {
+	// 	k.accumulationStore(ctx, coin.Denom).Increase(accumulationKey(lock.Duration), coin.Amount)
+	// }
 
 	k.hooks.OnTokenLocked(ctx, owner, lock.ID, lock.Coins, lock.Duration, lock.EndTime)
 	return nil
@@ -548,9 +549,9 @@ func (k Keeper) Unlock(ctx sdk.Context, lock types.PeriodLock) error {
 	}
 
 	// remove from accumulation store
-	for _, coin := range lock.Coins {
-		k.accumulationStore(ctx, coin.Denom).Decrease(accumulationKey(lock.Duration), coin.Amount)
-	}
+	// for _, coin := range lock.Coins {
+	// 	k.accumulationStore(ctx, coin.Denom).Decrease(accumulationKey(lock.Duration), coin.Amount)
+	// }
 
 	k.hooks.OnTokenUnlocked(ctx, owner, lock.ID, lock.Coins, lock.Duration, lock.EndTime)
 	return nil